[ORC] Introduce IRPartitionLayer for common partition functionality.
